+++ test/cbmc/proofs/MQTTPropGet_TopicAliasMax/MQTTPropGet_TopicAliasMax_harness.c
@@ -33,7 +33,7 @@ void harness()
 {
     MQTTPropBuilder_t * propBuffer;
     uint16_t * topicAliasMax;
-    uint32_t currentIndex;
+    size_t currentIndex;
 
     propBuffer = allocateMqttPropBuilder( NULL );
     __CPROVER_assume( isValidMqttPropBuilder( propBuffer ) );
